test utils, isApprox to expect_near
